src/requirements.txt
python-dotenv
sentence-transformers
faiss-cpu
numpy
tqdm
pymupdf
markdown
openai
streamlit
requests
